+++ example/chat_app/lib/services/audio_recording_service_stub.dart
@@ -31,6 +31,9 @@ class _UnsupportedAudioRecordingService implements AudioRecordingService {
   }
 
   @override
+  Future<Uint8List?> readUntrimmedRecording(String path) async => null;
+
+  @override
   Future<void> cancel() async {}
 
   @override